# ops/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Midnight API
  version: "0"
  description: |
    Instant REST API access to the Midnight blockchain.

    Query blocks, transactions, addresses, contracts, DUST generation and a
    raw ledger-event feed over plain REST. No API key, no signup required.

    Every successful response carries `results` plus the indexed chain tip at
    the time the request was served, so a client can tell how fresh a
    response is:

    ```json
    { "results": { "...": "endpoint-specific payload" }, "tip": { "height": 291741, "hash": "0x..." } }
    ```

    Paginated collection endpoints add a third field, `next_cursor`, present
    (a string) while there's a further page and `null` once exhausted; pass
    it back as the `cursor` query parameter to continue. Point lookups and
    unpaginated full-list endpoints never include `next_cursor` at all, there
    being no further page to fetch:

    ```json
    { "results": [{ "...": "..." }], "tip": { "height": 291741, "hash": "0x..." }, "next_cursor": "nf1...eyJ..." }
    ```

    Errors use a flat envelope instead:

    ```json
    { "status_code": 404, "error": "Not Found", "message": "transaction not found" }
    ```
  license:
    name: Apache-2.0
    url: https://github.com/mmahut/nightfrost/blob/main/LICENSE
security: []
servers:
  - url: https://mainnet.nightfrost.dev/api/v0
    description: Mainnet
  - url: https://preprod.nightfrost.dev/api/v0
    description: Preprod
  - url: https://preview.nightfrost.dev/api/v0
...
paths:
...
  /tx/submit:
    post:
      tags: [Submission]
      summary: Submit a signed transaction
      operationId: submitTransaction
      description: >
        The body is the raw ledger transaction, hex-encoded (optionally
        `0x`-prefixed) by default, or raw bytes when `Content-Type:
        application/octet-stream` is set. Proxied to the node as an unsigned
        `Midnight.send_mn_transaction` extrinsic; the transaction itself must
        already be fully signed and proven.
      requestBody:
        required: true
        content:
          text/plain:
            schema:
              type: string
              example: "0x1a2b3c..."
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: Accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  tx_hash:
                    type: string
                    description: 32-byte transaction hash, hex-encoded.
                required: [tx_hash]
        "400":
          description: Malformed body, or the node rejected the transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
...
  schemas:
    Error:
      type: object
      properties:
        status_code:
          type: integer
        error:
          type: string
        message:
          type: string
      required: [status_code, error, message]
...